added delete, logout, edit, design changes

+++ app/Http/Controllers/PostController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use App\Models\Post;
 
 class PostController extends Controller
@@ -30,7 +31,24 @@ class PostController extends Controller
 
     public function getDeletePost($post_id){
         $post=Post::where('id',$post_id)->first();
+        if(Auth::user() != $post->user){
+            return redirect()->back();
+        }
         $post->delete();
         return redirect()->route('dashboard')->with(['message'=>'Successfully Deleted!!']);
+    }
+    public function postEditPost(Request $request){
+        $request->validate([
+            'body'=>'required'
+        ],[
+            'body.required'=>'Please Enter Text'
+        ]);
+        $post=Post::find($request['postId']);
+        if(Auth::user() != $post->user){
+            return redirect()->back();
+        }
+        $post->body=$request['body'];
+        $post->update();
+        return response()->json(['new-body'=>$post->body],200);
     }
 }
